% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scART.R
\name{AddReadthroughAssay}
\alias{AddReadthroughAssay}
\title{Add readthrough assay}
\usage{
AddReadthroughAssay(
  seurat_object,
  data.dir,
  gene.column = 2,
  cell.column = 1,
  unique.features = TRUE,
  strip.suffix = FALSE
)
}
\arguments{
\item{seurat_object}{A seurat object}

\item{data.dir}{A directory}

\item{gene.column}{A number}

\item{cell.column}{A number}

\item{unique.features}{A boolean}

\item{strip.suffix}{A boolean}
}
\value{
A seurat object with a readthrough assay and overall cell readthrough level
}
\description{
Add readthrough assay
}
