% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scART.R
\name{GeneReadthrough}
\alias{GeneReadthrough}
\title{Calculate gene readthrough levels}
\usage{
GeneReadthrough(seurat_object, group.by, sort.by = mean)
}
\arguments{
\item{seurat_object}{A seurat object with a readthrough assay}

\item{group.by}{A grouping criterion}

\item{sort.by}{A sorting criterion}
}
\value{
sorted data frame
}
\description{
Calculate gene readthrough levels
}
