% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scART.R
\name{ReadthroughGTF}
\alias{ReadthroughGTF}
\title{Create GTF file annotating readthrough regions}
\usage{
ReadthroughGTF(
  input,
  output,
  chr.sizes = "",
  readthrough.gap = 2000,
  readthrough.window = 150000,
  min.window = 15000
)
}
\arguments{
\item{input}{A path to a gtf file}

\item{output}{A path to an output gtf file name}

\item{chr.sizes}{A path to a chromosome sizes file}

\item{readthrough.gap}{A number}

\item{readthrough.window}{A number}

\item{min.window}{A number}
}
\value{
void
}
\description{
Create GTF file annotating readthrough regions
}
